+++ bcbio/variation/samtools.py
@@ -37,9 +37,7 @@ def shared_variantcall(call_fn, name, align_bams, ref_file, items,
                         tx_out_file)
     if out_file.endswith(".gz"):
         out_file = vcfutils.bgzip_and_index(out_file, config)
-    ann_file = annotation.annotate_nongatk_vcf(out_file, align_bams, assoc_files.get("dbsnp"),
-                                               ref_file, items[0])
-    return ann_file
+    return out_file
 
 def run_samtools(align_bams, items, ref_file, assoc_files, region=None,
                  out_file=None):