+++ bcbio/variation/vardict.py
@@ -113,8 +113,7 @@ def _run_vardict_caller(align_bams, items, ref_file, assoc_files,
     if out_file is None:
         out_file = "%s-variants.vcf.gz" % os.path.splitext(align_bams[0])[0]
     if not utils.file_exists(out_file):
-        raw_file = "%s-raw%s" % utils.splitext_plus(out_file)
-        with file_transaction(items[0], raw_file) as tx_out_file:
+        with file_transaction(items[0], out_file) as tx_out_file:
             vrs = bedutils.population_variant_regions(items)
             target = shared.subset_variant_regions(
                 vrs, region, out_file, items=items, do_merge=False)
@@ -144,9 +143,9 @@ def _run_vardict_caller(align_bams, items, ref_file, assoc_files,
                        "| bcftools filter -i 'QUAL >= 0' "
                        "| {fix_ambig_ref} | {fix_ambig_alt} | {remove_dup} | {vcfstreamsort} {compress_cmd}")
                 if num_bams > 1:
-                    temp_file_prefix = raw_file.replace(".gz", "").replace(".vcf", "") + item["name"][1]
+                    temp_file_prefix = out_file.replace(".gz", "").replace(".vcf", "") + item["name"][1]
                     tmp_out = temp_file_prefix + ".temp.vcf"
-                    tmp_out += ".gz" if raw_file.endswith("gz") else ""
+                    tmp_out += ".gz" if out_file.endswith("gz") else ""
                     sample_vcf_names.append(tmp_out)
                     with file_transaction(item, tmp_out) as tx_tmp_file:
                         if not _is_bed_file(target):
@@ -166,10 +165,6 @@ def _run_vardict_caller(align_bams, items, ref_file, assoc_files,
                 vcfutils.merge_variant_files(orig_files=sample_vcf_names,
                                              out_file=tx_out_file, ref_file=ref_file,
                                              config=config, region=bamprep.region_to_gatk(region))
-        if assoc_files.get("dbsnp"):
-            annotation.add_dbsnp(raw_file, assoc_files["dbsnp"], items[0], out_file)
-        else:
-            utils.symlink_plus(raw_file, out_file)
     return out_file
 
 def depth_freq_filter(line, tumor_index, aligner):
@@ -239,8 +234,7 @@ def _run_vardict_paired(align_bams, items, ref_file, assoc_files,
     if out_file is None:
         out_file = "%s-paired-variants.vcf.gz" % os.path.splitext(align_bams[0])[0]
     if not utils.file_exists(out_file):
-        raw_file = "%s-raw%s" % utils.splitext_plus(out_file)
-        with file_transaction(items[0], raw_file) as tx_out_file:
+        with file_transaction(items[0], out_file) as tx_out_file:
             target = shared.subset_variant_regions(dd.get_variant_regions(items[0]), region,
                                                    out_file, do_merge=True)
             paired = vcfutils.get_paired_bams(align_bams, items)
@@ -292,10 +286,6 @@ def _run_vardict_paired(align_bams, items, ref_file, assoc_files,
                        "{somatic_filter} | {fix_ambig_ref} | {fix_ambig_alt} | {remove_dup} | {vcfstreamsort} "
                        "{compress_cmd} > {tx_out_file}")
                 do.run(cmd.format(**locals()), "Genotyping with VarDict: Inference", {})
-        if assoc_files.get("dbsnp"):
-            annotation.add_dbsnp(raw_file, assoc_files["dbsnp"], items[0], out_file)
-        else:
-            utils.symlink_plus(raw_file, out_file)
     return out_file
 
 def get_vardict_command(data):